requirements.txt
textblob==0.17.1
nltk==3.8.1
requests==2.31.0
snscrape==0.7.0
python-dotenv==1.0.1
pytest==8.2.1
